feat: change ids int size

+++ message.go
@@ -1,7 +1,7 @@
 package telegram
 
 type Message struct {
-	MessageID             int                   `json:"message_id"`
+	MessageID             int64                 `json:"message_id"`
 	From                  *User                 `json:"from,omitempty"`
 	Date                  int                   `json:"date"`
 	Chat                  *Chat                 `json:"chat"`
